saas/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bocai-backend"
version = "0.1.0"
description = "加拿大PC28自动投注托管SaaS平台后端"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic>=2.0",
    "aiosqlite",
    "aiohttp",
    "PyJWT",
    "python-multipart",
    "ddddocr",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
